Created a separate array of particle data to send the gpu to decrease bandwidth. Optimization.
Hopefully more performant than sending an array of the entire Particle struct as it keeps growing in size as I make the particle system more complicated. The data is only copied over at times when the particles are iterated anyway, so it should not be a performance loss on the cpu-side.